template: make sure that proper package manager is installed
before 02_install_groups step. Otherwise it may happen that package
specification (or some other behaviour) will be incompatible with
expectation of particular template version. This is all because Fedora
depending on version uses yum or dnf, which are only "almost"
compatible. For example yum needs environment groups with "@^" prefix,
while dnf with bare "@".