Don't set YUM_OPTS twice when we have a REPO_PROXY
prepare-chroot-builder is called through Makefile.fedora (which also
adds the proxy option) using sudo -E, so we end up adding the proxy twice,
once quoted and once not. This leads to errors like:

No package "--setopt=proxy=http://localhost:3128/" available.

When preparing the chroot. Fix it to not do that.