Workarounds for Debian's rpm configuration
On Debian, rpm is configured to keep rpmdb in ~/.rpmdb, even when
installing some system package with rpm --root=.... This makes it
problematic to hand it over to rpm running inside just installed system.
To solve this, override _dbpath setting. To do that, use bash as a
shell, for arrays.

QubesOS/qubes-issues#1907