Make build-deps step take signed src.rpm from another distribution version
`dnf build-deps` complains if the src.rpm is signed with an unknown key
(amusingly, it works just fine if the package isn't signed at all...).
In linux-dom0-updates repository we have packages backported from
another version, which is a problem here. `dnf build-deps` suggests
`--nogpgcheck` as the only solution, but it is wrong, because it will also
skip verification of packages downloaded from the network. Instead,
extract spec file manually and feed it into `dnf build-deps`.