generate-changelog: clamp timestamp to previous entry
Make sure dates in the changelog are in a chronological order, otherwise
rpm complains. Normally git log should sort entries just right, but if
there is an older commit with newer parent (for example result of a
rebase without resetting commit date), it will be wrong. Instead of
reordering the entries, adjust timestamps in that rare case. There is
commit ID anyway, so it is still possible to easily find the change.

While at it, do two more changes:
 - exclude merge commits
 - fix qubes-devel ML address