Fix handling dynamic src.rpm version with USE_DIST_BUILD_TOOLS=1
In case of dynamic src.rpm version (like with
INCREMENT_DEVEL_VERSIONS=1), mock will rebuild src.rpm and it will have
a different name (incremented version) than the original.
Don't fail at buildinfo generation step because of that - use wildcard
to catch the src.rpm. The buildinfo will still be named after original
src.rpm, but will also contain the new src.rpm name, so it will be
possible to match it to the package (and the information inside is
actually signed, contrary to the file name).