Forcibly rebuild rpmdb when constructing chroot
Initial packages are installed with rpm from the host system. This rpm
version may be hugely different than the one installed inside chroot
(and used for further actions). In most cases rpmdb --rebuildb is enough
to make rpmdb available to the in-chroot rpm, but in some cases
difference is that big that even this doesn't work. This is the case of
OpenSUSE Tumbleweed, where rpm switched from BDB to NDB database format,
which is completely unreadable by rpm package from Fedora.