Use dnf from the host for legacy builder too
Install packages in the build env using dnf from outside of the build
chroot. The main reason is to avoid outdated DNF (and its deps) in fc25
chroot (Qubes 4.0 dom0) talking to any network service. Old versions of
those tools are fine in the actual dom0 only because they are isolated
from the network - keep this property for the package building too.

Technically, it requires calling `dnf --installroot=...` from the
outside and adjust absolute paths (and file:// URLs) that used to be
valid inside chroot, to work now from the outside. In case of DNF
configuration, it requires introducing a variable with the chroot
location ($sysroot), otherwise file:// urls are incorrectly resolved.

The template build still uses dnf from inside. This is fine, because we
do not support EOL distributions in the templates. prepare-chroot-base
script is shared between those two cases. Since for the build env case
we keep using DNF (and rpm) from the outside, move rebuilding rpmdb for
in-chroot version to the template build script - to the place where we
switch to the in-chroot rpm version.

This also avoid the need to bind mount builder repository inside chroot.

And also switch all qubes repos to https URLs.