Define CentOS Stream as a separate dist
We tried so far to make a switch from CentOS 8 to CentOS Stream 8
transparent for the user but it leads to newer issues due to our
choice of using only python3.8 in CentOS 8. We decided to step back
on maintaining our own set of python3.8 packages (@fepitre's COPR
epel-8-python38 repository) due to huge amount of work that it
requires with respect to the litlle benefit we get (having Qubes PDF
converter notably).