fix: Allow signing *.buildinfo with SplitGPG setup
Issue: Signing *.buildinfo is hardcoded to `$(GNUPG)`.

In the SplitGPG setup we rely on the RPM macro `__gpg` to use the
`qubes-gpg-client-wrapper`. By default if the macro is not set rpm
returns (@least here on fc33) `gpg2`.

Solution: Relegate it to `rpm` like we do in the lines bellow for
signing packages.