Fix double parsing spec file
The legacy builder parses the spec file twice: once inside the chroot to
resolve all the macros in the correct environment, and then the second
time outside to install build deps. This mostly works, with exception to
the case of escaped '%' (as '%%') - if the spec is parsed twice, it gets
unescaped. Workaround this by double escaping before double parsing.

Fixes QubesOS/qubes-issues#6957