legacy: try harder escaping percent for build deps
If macro (included from another file) contains literal % that is
properly escaped, double parsing is going to mess it up. Fix it up by
escaping it back, so spec can be loaded to get BuildRequires headers. Do
not escape % as the first character on a line, as that's section header.
This doesn't produce fully correct spec yet, but should be enough to get
build deps. Actual build doesn't use double parsing, so it should work
just fine.