template: avoid ambiguity between updates and updates-testing
When updates-testing is enabled (between beta and final freeze) some
packages appear in both "updates" repo and "updates-testing", sometimes
with the same version. When that happens, it undefined which one will be
chosen by dnf. And that is problematic, because template build first
downloads packages, log their hashes and only then install them - with
--cacheonly option. If the other repository is chosen, such operation
fails, even though the package is already downloaded through the other
repo.

Fix this by changing updates-testing repo priority to be preferred one
(if enabled).

QubesOS/qubes-issues#9244